Convert tests to e2e tests
