--- test/e2e/tests/Unstable_TrapFocus/BaseTrapFocus.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import TrapFocus from '@material-ui/core/Unstable_TrapFocus';
-
-export default function BaseTrapFocus() {
-  return <div>Hello</div>;
-}
